add  LevelDone()
