+++ Assets/Scripts/ShakeCola.meta~HEAD
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9258ec5a6049f41fab4211b167257e3a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
