+++ Assets/Scripts/ShakeCola.meta~HEAD.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 94b661705b2c64a98b30daeaab8ca355
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
