+++ Assets/Scripts/ShakeCola.meta~HEAD_0.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: aa9cca425d46140bba03128e0338dd63
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
