+++ Assets/Scripts/ShakeCola.meta~e82d16d5b7533a3452952c0615d2dcbf60668fbe
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eec63a50750f8486a9d1cfe49f79a8df
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
