+++ Assets/Scripts/ShakeCola.meta~e82d16d5b7533a3452952c0615d2dcbf60668fbe.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b7c1cad7467f04de293cdfdba21d3ff9
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
