+++ Assets/Scripts/ShakeCola.meta~e82d16d5b7533a3452952c0615d2dcbf60668fbe_0.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0e5f5ec9a30604593be066973b5296f0
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
