+++ Assets/Scripts/ShakeCola/Endline.cs.meta~HEAD.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f6f92f4afa9c94ccb81a759250f54393
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
