+++ Assets/Scripts/ShakeCola/Endline.cs.meta~HEAD_0
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a505e4f3d6ea149f1b238ecd82aeb280
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
