+++ Assets/Scripts/ShakeCola/Endline.cs.meta~HEAD_0.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 02e715d46a60c45e8929d00c8f7754c9
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
