+++ Assets/Scripts/ShakeCola/Endline.cs.meta~e82d16d5b7533a3452952c0615d2dcbf60668fbe
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de5dfb94f7fd84132b2a6e4740c3fb08
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
