+++ Assets/Scripts/ShakeCola/Endline.cs.meta~e82d16d5b7533a3452952c0615d2dcbf60668fbe.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6a0a8576dc7ad49f8812a1d67a90d9db
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
