+++ Assets/Scripts/ShakeCola/Endline.cs.meta~e82d16d5b7533a3452952c0615d2dcbf60668fbe_0.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 591beca9382e34d1f83684c83da0f19a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
